Save 10/04 First version of an Auth

+++ frontend/src/main/frontend/src/app/library/library.details.component.ts
@@ -1,10 +1,9 @@
-import { Component, Input } from '@angular/core';
-import { Router } from '@angular/router'
+import { Component, Input, Output, EventEmitter  } from '@angular/core';
 import {Library} from "./library";
 import {LibrarySerivce} from "./library.service";
-import {PopoverModule} from "ngx-popover";
 import { DialogService } from "ng2-bootstrap-modal";
 import { ConfirmComponent } from '../confirm.component'
+import { NewLibraryFormComponent } from './new.library.form.component';
 
 
 @Component({
@@ -16,26 +15,49 @@ export class LibraryDetailsComponent {
   @Input()
   library: Library;
 
+  @Output() libraryDeleted = new EventEmitter();
+
 
   constructor(
     private _libraryService: LibrarySerivce,
-    private _router: Router,
     private dialogService:DialogService
   ) {}
 
   delete(){
     this._libraryService.delete(this.library)
       .subscribe(
-        data => this.library = data,
+        data => console.log(data),
         error => alert(error),
-        () => window.location.reload()
+        () => this.libraryDeleted.emit(this.library)
       );
-    window.location.reload();
   }
 
-  edit(){
+/*  edit(){
     this._libraryService.library = this.library;
     this._router.navigate(['/new-library']);
+  }*/
+
+  edit(){
+    this._libraryService.library = Object.assign({}, this.library);
+    this.dialogService.addDialog(NewLibraryFormComponent, {
+      title:'Edit of the Library: ',
+      library:this.library})
+      .subscribe((isConfirmed = true)=>{
+        if(isConfirmed){
+          this.newLibrary();
+        } else {
+          this.library = Object.assign({}, this._libraryService.library );
+
+        }
+      });
+  }
+
+  newLibrary(){
+    this._libraryService.create(this.library)
+      .subscribe(
+        data => this.library = data,
+        error => alert(error)
+      );
   }
 
   showConfirm() {
@@ -43,7 +65,9 @@ export class LibraryDetailsComponent {
       title:'Confirmation',
       message:'Are you sure you want to delete the library: ' + this.library.libraryName + ' ?'})
       .subscribe((isConfirmed)=>{
-        this.delete();
+        if(isConfirmed){
+          this.delete();
+        }
       });
   }
 }
